move db into api directory
